Refactor code and add method to load images as shared func
Mainly refactored the code so that they share the image loading part.
Facilitating debugging and implementing new models.
Also added a limit to channels read by the image loader as if the image
as 4 channels most model don't work.

There is a minor difference in output between this image loading and
SAM2 as I'm using cv2.resize to resize the image instead of PIL.resize.
The difference is minimal though so I'll keep cv2 as its more versatile.