Add basic support for DP3
As inital commit - I'm adding support for depth rendering using
DepthAnything3

To make it work with the structure of this repo - I had to edit some
files in their utils - and had to rename utils to dp_utils to avoid
conflicts with other utils folder in this repo. As a result - I had to
merge most of their code.